requirements.txt
numpy==1.26.4
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.9.0.80
matplotlib
pillow
